requirements.txt
websockets>=12.0
kafka-python>=2.0.2
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
mlflow>=2.10.0
evidently>=0.4.20
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=7.0.0
scipy>=1.11.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
locust>=2.20.0
prometheus-fastapi-instrumentator>=6.0.0
